完成vuex控制权限

+++ myweb-fe/src/components/admin/postsList.vue
@@ -57,6 +57,7 @@
 </template>
 <script>
 import {get, post} from '@/assets/js/util'
+import {mapGetters} from 'vuex'
 export default {
   name: 'postsList',
   data () {
@@ -69,16 +70,24 @@ export default {
   },
   methods: {
     handleEdit (index, row) {
-      this.$router.push({path: '/admin/editArticle', query: { id: row.id }})
+      if (this.isState) {
+        this.$router.push({path: '/admin/editArticle', query: { id: row.id }})
+      } else {
+        this.$message.error('权限不足')
+      }
     },
     handleDelete (index, row) {
-      this.$confirm('此操作将永久删除该文章, 是否继续?', '提示', {
-        confirmButtonText: '确定',
-        cancelButtonText: '取消',
-        type: 'warning'
-      }).then(() => {
-        this.deleteArticle('/api/posts/deleteArticle', {id: row.id})
-      })
+      if (this.isState) {
+        this.$confirm('此操作将永久删除该文章, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          this.deleteArticle('/api/posts/deleteArticle', {id: row.id})
+        })
+      } else {
+        this.$message.error('权限不足')
+      }
     },
     async deleteArticle (url, data) {
       const resMsg = await post(url, data)
@@ -113,6 +122,11 @@ export default {
   },
   created () {
     this.getList()
+  },
+  computed: {
+    ...mapGetters([
+      'isState'
+    ])
   }
 }
 </script>
